azure-functions/frete-bot/requirements.txt
azure-functions>=1.21.0
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
aiohttp>=3.9.0
